+++ product_portfolio/tests/__init__.py
@@ -18,6 +18,7 @@ from product_portfolio.models import ProductComponent
 from product_portfolio.models import ProductDependency
 from product_portfolio.models import ProductItemPurpose
 from product_portfolio.models import ProductPackage
+from product_portfolio.models import ProductStatus
 
 
 def make_product(dataspace, inventory=None, **data):
@@ -71,9 +72,24 @@ def make_product_component(product, component=None):
 
 
 def make_product_item_purpose(dataspace, **data):
+    if "label" not in data:
+        data["label"] = make_string(10)
+    if "text" not in data:
+        data["text"] = make_string(10)
+
     return ProductItemPurpose.objects.create(
-        label=make_string(10),
-        text=make_string(10),
+        dataspace=dataspace,
+        **data,
+    )
+
+
+def make_product_status(dataspace, **data):
+    if "label" not in data:
+        data["label"] = make_string(10)
+    if "text" not in data:
+        data["text"] = make_string(10)
+
+    return ProductStatus.objects.create(
         dataspace=dataspace,
         **data,
     )